refactor: Оптимизирован вывод <RoomScreen>
